- Changed yFormatter handling to be more flexible and produce more readable code for server pages. There are serverObject.putNum() methods to allow adding of number type values in a formatted form, and put() methods for number types that add them without formatting. This reduces the need to transform them into Strings in server pages and removes the HTML encoding step which is unecessary for numbers. - some minor code cleanups (mostly unnecessary casts, null checks)
git-svn-id: https://svn.berlios.de/svnroot/repos/yacy/trunk@4166 6c8d7289-2bf4-0310-a012-ef5d649a1542

+++ htroot/CrawlStartExpert_p.java
@@ -60,7 +60,7 @@ public class CrawlStartExpert_p {
             prop.put("crawlingIfOlderNumber", Math.round((float)crawlingIfOlder / (float)(60*24)));
             prop.put("crawlingIfOlderUnitDayCheck", 1);
         } else if (crawlingIfOlder >= 60) {
-            prop.put("crawlingIfOlderNumber", Math.round((float)crawlingIfOlder / 60f));
+            prop.put("crawlingIfOlderNumber", Math.round(crawlingIfOlder / 60f));
             prop.put("crawlingIfOlderUnitHourCheck", 1);
         } else {
             prop.put("crawlingIfOlderNumber", crawlingIfOlder);
